+++ src/App.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import logo from './logo.svg';
 import './App.css';
 import Header from './Header';
-import Main from './Main';
 import Footer from './Footer';
+import Main from './마이페이지/Main';
 
 function App() {
   return (
