+++ src/마이페이지/WriteReview.js
@@ -1,0 +1,6 @@
+function WriteReview() {
+  console.log("리뷰버튼");
+  return <button className="writeReview-btn">리뷰 작성하기</button>;
+}
+
+export default WriteReview;
